requirements.txt
fastapi
uvicorn
pydub
google-cloud-speech
librosa
numpy